Redo index storage and only suggest words from same window
Instead of reacting `on_modified_async` use `view.change_count()` to
decide if a view needs to be re-indexed.  Register `on_close` to cleanup
the index.

For now select the other views from the same window.  But really, as
the approach is rather flexible this can be changed more easily.