Install low priority package for the key binding
As we bind the `tab` key wit low specificity, we interfere with other
packages, for example "Terminus" and "GitSavvy", which also register
`tab` but usually with very high specificity.  We don't want to special
case all probable plugins here but rather install the key binding with
a low priority.

For this to work we just need to create a zipped package with *just*
the key binding and a special name because the packages load sorted
by name.